test: create utils for method useful in tests
